+++ src/plugins/data_platform/project.py
@@ -2,10 +2,11 @@ import requests
 import json
 import math
 import pandas as pd
-import plotly
-import plotly.express as px
-import plotly.graph_objects as go
+import pytz
+import time
 from datetime import datetime
+from irods.session import iRODSSession
+from irods.models import RuleExec
 from flask import (
     Blueprint,
     render_template,
@@ -14,13 +15,26 @@ from flask import (
     url_for,
     session,
     flash,
+    current_app,
 )
-
+from cache import cache
+from signals import mango_signals
+from csrf import csrf
 from . import API_URL, current_user_api_token, openid_login_required, Session
 
 data_platform_project_bp = Blueprint(
     "data_platform_project_bp", __name__, template_folder="templates"
 )
+
+project_changed = mango_signals.signal("project_changed")
+
+
+def project_user_search_cache_update_listener(sender, **params):
+    cache_item_path = f"view/{url_for('data_platform_project_bp.project_user_search')}"
+    cache.delete(cache_item_path)
+
+
+project_changed.connect(project_user_search_cache_update_listener)
 
 
 @data_platform_project_bp.route(
@@ -55,6 +69,13 @@ def project(project_name):
 
     status = response.json()
 
+    response = requests.get(
+        f"{API_URL}/v1/projects/{project_name}/quota", headers=header
+    )
+    response.raise_for_status()
+
+    quotalog = response.json()
+
     project["activated"] = (
         not project["archived"]
         or not project["valid_after"]
@@ -86,6 +107,7 @@ def project(project_name):
         "project/project_view.html.j2",
         project=project,
         status=status,
+        quotalog=quotalog,
         zones=zones,
         admin=("operator" in perms or "admin" in perms),
     )
@@ -112,6 +134,8 @@ def add_project_member():
 
     flash(response.json()["message"], "success")
 
+    project_changed.send(current_app._get_current_object())
+
     return redirect(url_for("data_platform_project_bp.project", project_name=id))
 
 
@@ -133,6 +157,8 @@ def delete_project_member():
 
     flash(response.json()["message"], "success")
 
+    project_changed.send(current_app._get_current_object())
+
     return redirect(url_for("data_platform_project_bp.project", project_name=id))
 
 
@@ -146,8 +172,10 @@ def modify_project():
 
     if "description" in request.form:
         data = {
+            "type": request.form.get("type"),
             "description": request.form.get("description"),
             "sap_ref": request.form.get("sap_ref"),
+            "an": request.form.get("an"),
             "vsc_call": request.form.get("vsc_call"),
             "valid_after": request.form.get("valid_after"),
             "invalid_after": request.form.get("invalid_after"),
@@ -183,6 +211,32 @@ def modify_project():
 
     flash(response.json()["message"], "success")
 
+    project_changed.send(current_app._get_current_object())
+
+    return redirect(url_for("data_platform_project_bp.project", project_name=id))
+
+@data_platform_project_bp.route("/data-platform/projects/modify/rdr", methods=["POST"])
+@openid_login_required
+def modify_project_rdr():
+    token, _ = current_user_api_token()
+    header = {"Authorization": "Bearer " + token}
+
+    id = request.form.get("project")
+
+    for key in ['s3_bucket', 's3_prefix']:
+        value = request.form.get(key, "")
+
+        response = requests.put(
+            f"{API_URL}/v1/projects/{id}/option/{key}",
+            headers=header,
+            json={"value": value},
+        )
+        response.raise_for_status()
+
+        flash(response.json()["message"], "success")
+
+        project_changed.send(current_app._get_current_object())
+
     return redirect(url_for("data_platform_project_bp.project", project_name=id))
 
 
@@ -205,6 +259,7 @@ def set_project_options():
         options += [
             "sftp-openfile",
             "enable-icommands",
+            "enable-sftp-ingress",
             "enforce-quota",
             "inherit-permissions",
             "strict-permissions",
@@ -230,6 +285,8 @@ def set_project_options():
         response.raise_for_status()
 
         flash(response.json()["message"], "success")
+
+        project_changed.send(current_app._get_current_object())
 
     return redirect(url_for("data_platform_project_bp.project", project_name=id))
 
@@ -277,6 +334,8 @@ def deploy_project():
     response.raise_for_status()
 
     flash(response.json()["message"], "success")
+
+    project_changed.send(current_app._get_current_object())
 
     return redirect(url_for("data_platform_project_bp.project", project_name=id))
 
@@ -339,22 +398,20 @@ def add_irods_project():
         f"{API_URL}/v1/projects/{id}",
         headers=header,
         json={
-            "platform": request.form.get("platform"),
+            "type": request.form.get("type"),
+            "platform": "irods",
             "platform_options": [
                 {
                     "key": "zone-jobid",
                     "value": request.form.get("zone"),
                 },
-                {
-                    "key": "folder-layout",
-                    "value": request.form.get("layout"),
-                },
             ],
         },
     )
     response.raise_for_status()
-
     flash(response.json()["message"], "success")
+
+    project_changed.send(current_app._get_current_object())
 
     return redirect(url_for("data_platform_project_bp.project", project_name=id))
 
@@ -371,6 +428,7 @@ def add_generic_project():
         f"{API_URL}/v1/projects/{id}",
         headers=header,
         json={
+            "type": request.form.get("type"),
             "platform": "generic",
         },
     )
@@ -378,8 +436,42 @@ def add_generic_project():
 
     flash(response.json()["message"], "success")
 
+    project_changed.send(current_app._get_current_object())
+
     return redirect(url_for("data_platform_project_bp.project", project_name=id))
 
+@data_platform_project_bp.route("/data-platform/projects/add/rdr", methods=["POST"])
+@openid_login_required
+def add_rdr_project():
+    token, _ = current_user_api_token()
+    header = {"Authorization": "Bearer " + token}
+
+    id = request.form.get("name")
+
+    response = requests.put(
+        f"{API_URL}/v1/projects/{id}",
+        headers=header,
+        json={
+            "platform": "rdr",
+            "platform_options": [
+                {
+                    "key": "s3_bucket",
+                    "value": request.form.get("s3_bucket"),
+                },
+                {
+                    "key": "s3_prefix",
+                    "value": request.form.get("s3_prefix"),
+                },
+            ],
+        },
+    )
+    response.raise_for_status()
+
+    flash(response.json()["message"], "success")
+
+    project_changed.send(current_app._get_current_object())
+
+    return redirect(url_for("data_platform_project_bp.project", project_name=id))
 
 @data_platform_project_bp.route("/data-platform/projects", methods=["GET"])
 @openid_login_required
@@ -426,6 +518,15 @@ def convert_bytes_to_GB(size_bytes, conversion_to="GB"):
     return round(float_size, 2)
 
 
+def calculate_usage_percent(quota, usage):
+    if quota == 0:
+        return "No quota set!"
+    if quota > 0 and usage == 0:
+        return "0 %"
+    usage_percent = (usage / quota) * 100
+    return f"{round(usage_percent, 2)} %"
+
+
 @data_platform_project_bp.route("/data-platform/statistics", methods=["GET"])
 @openid_login_required
 def projects_statistics():
@@ -448,18 +549,26 @@ def projects_statistics():
         projects = []
 
     def create_project_dict(project):
-        zone_name = (
-            "-".join(project["project"]["platform_options"][0]["value"].split("-")[4:])
-            if project["project"]["platform"] == "irods"
-            else "Non iRODS"
-        )
+        if project["project"]["platform"] == "irods":
+            zone_name = [
+                "-".join(x["value"].split("-")[4:])
+                for x in project["project"]["platform_options"]
+                if x["key"] == "zone-jobid"
+            ][0]
+        else:
+            zone_name = "Non iRODS"
         return {
             "zone_name": zone_name,
             "project_name": project["project"]["name"],
+            "project_type": project["project"]["type"],
             "usage_total": convert_bytes_to_GB(
                 [x["used_size"] for x in project["usage"]][-1]
             ),
             "quota_set": convert_bytes_to_GB(project["project"]["quota_size"]),
+            "quota_usage_rate": calculate_usage_percent(
+                project["project"]["quota_size"],
+                [x["used_size"] for x in project["usage"]][-1],
+            ),
             "responsible_name": project["responsibles"][0]["name"]
             if project["responsibles"] != None
             else "",
@@ -478,26 +587,61 @@ def projects_statistics():
     )
 
 
-@data_platform_project_bp.route("/data-platform/statistics/usage", methods=["GET"])
+def summarize(source_name: str, data: pd.DataFrame, y_axis: str, group_index: int, hovertemplate=""):
+    def get_color(i):
+        color_mapping = [
+            '#4e79a7',
+            '#f28e2c',
+            '#e15759',
+            '#76b7b2',
+            '#59a14f',
+            '#edc949',
+            '#af7aa1',
+            '#ff9da7',
+            '#9c755f',
+            '#bab0ab'
+        ]
+        return color_mapping[i]
+
+    grouped_data = data.groupby("date")
+    if y_axis == "usage":
+        y = list(grouped_data.usage.agg("sum"))
+    elif y_axis == "quota":
+        y = list(grouped_data.quota.agg("sum"))
+    else:
+        raise ValueError("should be usage or quota")
+
+    group_data = {
+        "x": [month for month, _ in grouped_data],
+        "y": y,
+        "name": source_name,
+        "marker": {"color": get_color(group_index)},
+        "type": "bar",
+    }
+    if hovertemplate:
+        group_data["hovertemplate"] = hovertemplate
+    return group_data
+
+
+def get_next_month(year_month: str):
+    year, month = year_month.split("-")
+    if month == "12":
+        return f"{int(year)+1}-01"
+    else:
+        return f"{year}-{int(month)+1:02d}"
+
+
+@data_platform_project_bp.route("/data-platform/statistics/usage", methods=["GET", "POST"])
 @openid_login_required
+@csrf.exempt
 def projects_usage():
     token, _ = current_user_api_token()
     header = {"Authorization": "Bearer " + token}
 
-    year = request.args.get("year")
-
-    if not year:
-        year = datetime.now().year
-
-    response = requests.get(f"{API_URL}/v1/projects/usage/{year}", headers=header)
-
-    response.raise_for_status()
-
-    projects = response.json()
-
-    if not projects:
-        flash(f"No project information found in {year}.")
-        projects = []
+    start_date = "2023-01"
+    end_date = str(time.strftime("%Y-%m"))
+    start_date_year = int(start_date.split("-")[0])
+    end_date_year = int(end_date.split("-")[0])
 
     projects_dict = {}
     projects_dict["date"] = []
@@ -505,60 +649,217 @@ def projects_usage():
     projects_dict["project_name"] = []
     projects_dict["usage"] = []
     projects_dict["quota"] = []
-    for project in projects:
-        if project["project"]["platform"] == "irods":
-            for usage in project["usage"]:
-                projects_dict["date"].append(usage["date"])
-                zone_name = "-".join(
-                    project["project"]["platform_options"][0]["value"].split("-")[4:]
-                )
-                projects_dict["zone"].append(zone_name)
-                projects_dict["project_name"].append(project["project"]["name"])
-                projects_dict["usage"].append(convert_bytes_to_GB(usage["used_size"]))
-                projects_dict["quota"].append(convert_bytes_to_GB(usage["quota_size"]))
+    for year in range(start_date_year, end_date_year+1):
+        response = requests.get(f"{API_URL}/v1/projects/usage/{year}", headers=header)
+        response.raise_for_status()
+        projects = response.json()
+        for project in projects:
+            if project["project"]["platform"] == "irods" and project["project"]["type"] == "project":
+                for usage in project["usage"]:
+                    projects_dict["date"].append(usage["date"])
+                    zone_name = [
+                        "-".join(x["value"].split("-")[4:])
+                        for x in project["project"]["platform_options"]
+                        if x["key"] == "zone-jobid"
+                    ][0]
+                    projects_dict["zone"].append(zone_name)
+                    projects_dict["project_name"].append(project["project"]["name"])
+                    projects_dict["usage"].append(usage["used_size"])
+                    projects_dict["quota"].append(usage["quota_size"])
 
-    df_raw = pd.DataFrame(projects_dict)
-    df = (
-        df_raw.groupby(["date", "zone", "quota", "project_name"])["usage"]
-        .sum()
-        .reset_index(name="used_size")
-    )
+    df = pd.DataFrame(projects_dict)
 
-    fig_usage = px.histogram(
-        df, x="date", y=df["used_size"], color="zone", barmode="stack", text_auto=True
-    )
-    fig_usage.update_layout(
-        title="Usage Per Zone",
-        title_x=0.5,
-        autosize=True,
-        margin=dict(
-            autoexpand=True,
-            l=100,
-            r=20,
-            t=110,
-        ),
-        plot_bgcolor="white",
-    )
+    filters = {
+        "start_date": start_date,
+        "end_date": end_date,
+        "from_date": start_date,
+        "to_date": end_date,
+    }
 
-    fig_quota = px.histogram(
-        df, x="date", y=df["quota"], color="zone", barmode="stack", text_auto=True
-    )
-    fig_quota.update_layout(
-        title="Quota Per Zone",
-        title_x=0.5,
-        autosize=True,
-        margin=dict(
-            autoexpand=True,
-            l=100,
-            r=20,
-            t=110,
-        ),
-        plot_bgcolor="white",
-    )
+    if request.method == "POST":
+        filters["from_date"] = request.form.get("from-month", filters["from_date"])
+        filters["to_date"] = request.form.get("to-month", filters["to_date"])
+        df = df.loc[
+            (df.date >= filters["from_date"])
+            & (df.date < get_next_month(filters["to_date"]))
+        ]
+
+    all_groups = list(df.zone.unique())
+    all_groups.sort()
+
+    usage_plot = [
+        summarize(
+            source_name,
+            data,
+            "usage",
+            all_groups.index(source_name),
+            "%{y:.4s}B data used in %{x}",
+        )
+        for source_name, data in df.groupby("zone")
+    ]
+
+    quota_plot = [
+        summarize(
+            source_name,
+            data,
+            "quota",
+            all_groups.index(source_name),
+            "%{y:.4s}B quota set in %{x}",
+        )
+        for source_name, data in df.groupby("zone")
+    ]
 
     return render_template(
         "project/projects_usage.html.j2",
-        year=year,
-        usage_graphJSON=json.dumps(fig_usage, cls=plotly.utils.PlotlyJSONEncoder),
-        quota_graphJSON=json.dumps(fig_quota, cls=plotly.utils.PlotlyJSONEncoder),
+        usage_plot=usage_plot,
+        quota_plot=quota_plot,
+        filters=filters,
+    )
+
+
+@data_platform_project_bp.route("/data-platform/project_user_search", methods=["GET"])
+@openid_login_required
+@cache.cached(timeout=3600)
+def project_user_search():
+    token, _ = current_user_api_token()
+    header = {"Authorization": "Bearer " + token}
+
+    response = requests.get(f"{API_URL}/v1/projects", headers=header)
+    response.raise_for_status()
+
+    projects = response.json()
+
+    projects_list = []
+    for project in projects:
+        if project["platform"] == "irods":
+            zone_name = [
+                "-".join(x["value"].split("-")[4:])
+                for x in project["platform_options"]
+                if x["key"] == "zone-jobid"
+            ][0]
+            projects_list.append((zone_name, project["name"], project["type"]))
+        else:
+            projects_list.append(("Non iRODS", project["name"], ""))
+    project_list_of_dicts = []
+    for project in projects_list:
+        response = requests.get(
+            f"{API_URL}/v1/projects/{project[1]}/members", headers=header
+        )
+        members = response.json()
+        for member in members:
+            project_list_of_dicts.append(
+                {
+                    "user_name": member["name"],
+                    "user_account": member["username"],
+                    "user_role": member["role"],
+                    "user_email": member["email"],
+                    "project_name": project[1],
+                    "project_type": project[2],
+                    "zone_name": project[0],
+                }
+            )
+
+    return render_template(
+        "project/project_user_search.html.j2",
+        user_project_search_list=json.dumps(project_list_of_dicts),
+    )
+
+
+@data_platform_project_bp.route("/data-platform/rule-management", methods=["GET"])
+@openid_login_required
+def rule_management():
+    token, _ = current_user_api_token()
+    header = {"Authorization": "Bearer " + token}
+
+    def get_zones():
+        response = requests.get(f"{API_URL}/v1/irods/zones", headers=header)
+        response.raise_for_status()
+        response = response.json()
+        return [item["jobid"] for item in response]
+
+    def get_irods_credentials(jobid):
+        response = requests.post(
+            f"{API_URL}/v1/irods/zones/{jobid}/admin_token", headers=header
+        )
+        response.raise_for_status()
+        response = response.json()
+        irods_environment = response["irods_environment"]
+        password = response["token"]
+        return irods_environment, password
+
+    rule_info = []
+    for zone in get_zones():
+        zone_environment, password = get_irods_credentials(zone)
+        with iRODSSession(**zone_environment, password=password) as session:
+            query = session.query(
+                RuleExec.name,
+                RuleExec.id,
+                RuleExec.user_name,
+                RuleExec.time,
+                RuleExec.last_exe_time,
+                RuleExec.frequency,
+            )
+            for item in query:
+                rule_info.append(list(item.values()))
+                rule_info[-1].insert(0, session.zone)
+
+    def localize_datetime(
+        value, format="%Y-%m-%d %H:%M:%S", local_timezone="Europe/Brussels"
+    ):
+        tz = pytz.timezone(local_timezone)
+        utc = pytz.timezone("UTC")
+        value = utc.localize(value, is_dst=None).astimezone(pytz.utc)
+        local_dt = value.astimezone(tz)
+        local_dt = local_dt.strftime(format)
+        return datetime.strptime(local_dt, format)
+
+    for item in rule_info:
+        if item[5] is not None:
+            item[4] = localize_datetime(item[4])
+            item[5] = localize_datetime(item[5])
+            delta = item[4] - datetime.now().replace(microsecond=0)
+            item.insert(4, delta)
+        else:
+            item.insert(4, "Not executed yet!")
+
+    return render_template(
+        "project/rule_management.html.j2",
+        rule_info=rule_info,
+    )
+
+
+@data_platform_project_bp.route("/data-platform/quota-change", methods=["GET"])
+@openid_login_required
+def project_quota_change():
+    token, _ = current_user_api_token()
+    header = {"Authorization": "Bearer " + token}
+
+    response = requests.get(f"{API_URL}/v1/projects/quota", headers=header)
+
+    response.raise_for_status()
+
+    projects = response.json()
+
+    projects_list = []
+    for project in projects:
+        for day in project["log"]:
+            projects_list.append(
+                {
+                    "project_name": project["name"],
+                    "project_create": project["log"][0]["date"],
+                    "project_type": project["type"],
+                    "project_status": (f"Active" if day["archived"]==False else "Archived"),
+                    "sap_ref": project["sap_ref"],
+                    "an": project["an"],
+                    "quota_set": convert_bytes_to_GB(
+                        day["quota_size"], conversion_to="TB"
+                    ) if day["archived"]==False else None,
+                    "quota_set_date": None if (day["quota_size"]==0 or day["archived"]==True) else day["date"],
+                    "modified_by": day["modified_by"],
+                }
+            )
+
+    return render_template(
+        "project/projects_quota_change.html.j2",
+        projects_list=json.dumps(projects_list),
     )