+++ src/plugins/mango_open_search/search.py
@@ -37,6 +37,16 @@ def index_collection():
         # select the second element because the first one is empty
         zone = collection_path.split("/")[1]
 
+    # delete all subtree items from the index below the requested path
+
+    if "skip_reset_collection" not in request.form:
+        add_index_job(
+            zone=zone,
+            job_type="delete_subtree",
+            item_path=collection_path,
+            item_type="collection",
+        )
+
     add_index_job(
         zone=zone,
         job_type="index_item",
@@ -69,6 +79,10 @@ def zone_search():
     search_string = ""
     if "search_string" in request.values and request.values["search_string"]:
         search_string = request.values["search_string"]
+        # combined user and group ids into a global list
+        irods_acl_reader_ids = [g.irods_session.user.id]
+        irods_acl_reader_ids.extend(g.irods_session.my_group_ids)
+
         search_results = mango_osc.search(
             {
                 "query": {
@@ -77,7 +91,13 @@ def zone_search():
                             {
                                 "multi_match": {
                                     "query": request.values["search_string"],
-                                    "fields": ["match_all"],
+                                    "fields": [
+                                        "irods_name^4",
+                                        "mango_descriptive_text_basket",
+                                    ],
+                                    "fuzziness": "AUTO",
+                                    "fuzzy_transpositions": True,
+                                    "minimum_should_match": 1,
                                 }
                             },
                             {
@@ -89,22 +109,8 @@ def zone_search():
                                             }
                                         },
                                         {
-                                            "bool": {
-                                                "should": [
-                                                    {
-                                                        "term": {
-                                                            "acl_read_users": g.irods_session.user.id
-                                                        }
-                                                    }
-                                                ]
-                                                + [
-                                                    {
-                                                        "term": {
-                                                            "acl_read_groups": group_id
-                                                        }
-                                                    }
-                                                    for group_id in g.irods_session.my_group_ids
-                                                ]
+                                            "terms": {
+                                                "irods_acl_reader_ids": irods_acl_reader_ids
                                             }
                                         },
                                     ]